+++ apps/commerce-streamer/src/main/kotlin/com/loopers/domain/order/dto/OrderCompletedMessage.kt
@@ -1,4 +1,4 @@
-package com.loopers.domain.order
+package com.loopers.domain.order.dto
 
 data class OrderCompletedMessage(
     val orderId: Long,
